LayoutTests/media/session/session-creation-expected.txt

New media sessions should be in an Idle state after creation.


---
Testing "new MediaSession('content')"
EXPECTED (internals.mediaSessionCurrentState(session) == 'idle') OK
---

---
Testing "new MediaSession('transient')"
EXPECTED (internals.mediaSessionCurrentState(session) == 'idle') OK
---

---
Testing "new MediaSession('transient-solo')"
EXPECTED (internals.mediaSessionCurrentState(session) == 'idle') OK
---

---
Testing "new MediaSession('ambient')"
EXPECTED (internals.mediaSessionCurrentState(session) == 'idle') OK
---

---
Testing "new MediaSession()"
EXPECTED (internals.mediaSessionCurrentState(session) == 'idle') OK
---

END OF TEST
